Implement more priors, add tests, use throughout package
