+++ gpytorch/kernels/linear_kernel.py
@@ -6,7 +6,6 @@ from __future__ import unicode_literals
 import torch
 from gpytorch.kernels import Kernel
 from gpytorch.lazy import MatmulLazyTensor, RootLazyTensor
-from gpytorch.priors._compatibility import _bounds_to_prior
 
 
 class LinearKernel(Kernel):
@@ -58,9 +57,7 @@ class LinearKernel(Kernel):
         offset_bounds=None,
     ):
         super(LinearKernel, self).__init__(active_dims=active_dims)
-        variance_prior = _bounds_to_prior(prior=variance_prior, bounds=variance_bounds, log_transform=False)
         self.register_parameter(name="variance", parameter=torch.nn.Parameter(torch.zeros(1)), prior=variance_prior)
-        offset_prior = _bounds_to_prior(prior=offset_prior, bounds=offset_bounds, log_transform=False)
         self.register_parameter(
             name="offset", parameter=torch.nn.Parameter(torch.zeros(1, 1, num_dimensions)), prior=offset_prior
         )
