+++ gpytorch/kernels/matern_kernel.py
@@ -85,7 +85,6 @@ class MaternKernel(Kernel):
         active_dims=None,
         eps=1e-6,
         log_lengthscale_prior=None,
-        log_lengthscale_bounds=None,
     ):
         if nu not in {0.5, 1.5, 2.5}:
             raise RuntimeError("nu expected to be 0.5, 1.5, or 2.5")
@@ -95,7 +94,6 @@ class MaternKernel(Kernel):
             log_lengthscale_prior=log_lengthscale_prior,
             active_dims=active_dims,
             batch_size=batch_size,
-            log_lengthscale_bounds=log_lengthscale_bounds,
         )
         self.nu = nu
         self.eps = eps
