+++ gpytorch/kernels/periodic_kernel.py
@@ -6,7 +6,6 @@ from __future__ import unicode_literals
 import math
 import torch
 from .kernel import Kernel
-from ..priors._compatibility import _bounds_to_prior
 
 
 class PeriodicKernel(Kernel):
@@ -79,16 +78,12 @@ class PeriodicKernel(Kernel):
         eps=1e-6,
         log_lengthscale_prior=None,
         log_period_length_prior=None,
-        log_lengthscale_bounds=None,
-        log_period_length_bounds=None,
     ):
-        log_period_length_prior = _bounds_to_prior(prior=log_period_length_prior, bounds=log_period_length_bounds)
         super(PeriodicKernel, self).__init__(
             has_lengthscale=True,
             active_dims=active_dims,
             batch_size=batch_size,
             log_lengthscale_prior=log_lengthscale_prior,
-            log_lengthscale_bounds=log_lengthscale_bounds,
         )
         self.eps = eps
         self.register_parameter(
