+++ gpytorch/likelihoods/gaussian_likelihood.py
@@ -7,7 +7,6 @@ import math
 import torch
 from gpytorch.functions import add_diag
 from gpytorch.likelihoods import Likelihood
-from gpytorch.priors._compatibility import _bounds_to_prior
 from gpytorch.random_variables import GaussianRandomVariable
 
 
@@ -15,11 +14,9 @@ class GaussianLikelihood(Likelihood):
     """
     """
 
-    def __init__(self, log_noise_prior=None, log_noise_bounds=None):
-        # TODO: Remove deprecated log_noise_bounds kwarg
-        log_noise_prior = _bounds_to_prior(prior=log_noise_prior, bounds=log_noise_bounds)
+    def __init__(self, log_noise_prior=None):
         super(GaussianLikelihood, self).__init__()
-        self.register_parameter(name="log_noise", parameter=torch.nn.Parameter(torch.tensor(0.)), prior=log_noise_prior)
+        self.register_parameter(name="log_noise", parameter=torch.nn.Parameter(torch.tensor([0.])), prior=log_noise_prior)
 
     def forward(self, input):
         assert isinstance(input, GaussianRandomVariable)
