+++ gpytorch/means/constant_mean.py
@@ -5,13 +5,10 @@ from __future__ import unicode_literals
 
 import torch
 from gpytorch.means import Mean
-from gpytorch.priors._compatibility import _bounds_to_prior
 
 
 class ConstantMean(Mean):
-    def __init__(self, prior=None, batch_size=None, constant_bounds=None):
-        # TODO: Remove deprecated bounds kwarg
-        prior = _bounds_to_prior(prior=prior, bounds=constant_bounds, batch_size=batch_size, log_transform=False)
+    def __init__(self, prior=None, batch_size=None):
         super(ConstantMean, self).__init__()
         self.batch_size = batch_size
         self.register_parameter(
