--- gpytorch/priors_torch/__init__.py
@@ -1,19 +1,0 @@
-from __future__ import absolute_import, division, print_function, unicode_literals
-
-from .lkj_prior import LKJCholeskyFactorPrior, LKJCovariancePrior, LKJPrior
-# from .smoothed_box_prior import SmoothedBoxPrior
-from .torch_priors import GammaPrior, MultivariateNormalPrior, NormalPrior
-
-
-# from .wishart_prior import InverseWishartPrior, WishartPrior
-
-
-__all__ = [
-    "GammaPrior",
-    "LKJPrior",
-    "LKJCholeskyFactorPrior",
-    "LKJCovariancePrior",
-    "MultivariateNormalPrior",
-    "NormalPrior",
-    # "SmoothedBoxPrior",
-]
