+++ test/examples/test_hadamard_multitask_gp_regression.py
@@ -1,20 +1,19 @@
-from __future__ import absolute_import
-from __future__ import division
-from __future__ import print_function
-from __future__ import unicode_literals
+from __future__ import absolute_import, division, print_function, unicode_literals
 
 import os
 import random
-import torch
 import unittest
+from math import exp, pi
+
 import gpytorch
-from torch import optim
-from gpytorch.kernels import RBFKernel, IndexKernel
+import torch
+from gpytorch.kernels import IndexKernel, RBFKernel
 from gpytorch.likelihoods import GaussianLikelihood
 from gpytorch.means import ConstantMean
-from gpytorch.priors import InverseWishartPrior, SmoothedBoxPrior
+from gpytorch.priors import LKJCovariancePrior, SmoothedBoxPrior
 from gpytorch.random_variables import GaussianRandomVariable
-from math import pi
+from torch import optim
+
 
 # Simple training data: let's try to learn a sine function
 train_x = torch.linspace(0, 1, 100)
@@ -24,8 +23,8 @@ train_y1 = torch.sin(train_x * (2 * pi)) + torch.randn_like(train_x).mul_(1e-2)
 train_y2 = torch.cos(train_x * (2 * pi)) + torch.randn_like(train_x).mul_(1e-2)
 
 test_x = torch.linspace(0, 1, 51)
-y1_inds_test = torch.zeros(51).long()
-y2_inds_test = torch.ones(51).long()
+y1_inds_test = torch.zeros(51, dtype=torch.long)
+y2_inds_test = torch.ones(51, dtype=torch.long)
 test_y1 = torch.sin(test_x * (2 * pi))
 test_y2 = torch.cos(test_x * (2 * pi))
 
@@ -39,7 +38,9 @@ class HadamardMultitaskGPModel(gpytorch.models.ExactGP):
         self.covar_module = RBFKernel()
         # We learn an IndexKernel for 2 tasks
         # (so we'll actually learn 2x2=4 tasks with correlations)
-        self.task_covar_module = IndexKernel(n_tasks=2, rank=1, prior=InverseWishartPrior(nu=2, K=torch.eye(2)))
+        sd_prior = SmoothedBoxPrior(exp(-4), exp(4), log_transform=True)
+        cov_prior = LKJCovariancePrior(n=2, eta=1, sd_prior=sd_prior)
+        self.task_covar_module = IndexKernel(n_tasks=2, rank=1, prior=cov_prior)
 
     def forward(self, x, i):
         # Get predictive mean
@@ -54,7 +55,10 @@ class HadamardMultitaskGPModel(gpytorch.models.ExactGP):
 
 class TestHadamardMultitaskGPRegression(unittest.TestCase):
     def setUp(self):
-        if os.getenv("UNLOCK_SEED") is None or os.getenv("UNLOCK_SEED").lower() == "false":
+        if (
+            os.getenv("UNLOCK_SEED") is None
+            or os.getenv("UNLOCK_SEED").lower() == "false"
+        ):
             self.rng_state = torch.get_rng_state()
             torch.manual_seed(0)
             if torch.cuda.is_available():
@@ -68,7 +72,9 @@ class TestHadamardMultitaskGPRegression(unittest.TestCase):
     def test_multitask_gp_mean_abs_error(self):
         likelihood = GaussianLikelihood(log_noise_prior=SmoothedBoxPrior(-6, 6))
         gp_model = HadamardMultitaskGPModel(
-            (torch.cat([train_x, train_x]), torch.cat([y1_inds, y2_inds])), torch.cat([train_y1, train_y2]), likelihood
+            (torch.cat([train_x, train_x]), torch.cat([y1_inds, y2_inds])),
+            torch.cat([train_y1, train_y2]),
+            likelihood,
         )
         mll = gpytorch.mlls.ExactMarginalLogLikelihood(likelihood, gp_model)
 
@@ -78,7 +84,9 @@ class TestHadamardMultitaskGPRegression(unittest.TestCase):
         optimizer = optim.Adam(gp_model.parameters(), lr=0.01)
         for _ in range(100):
             optimizer.zero_grad()
-            output = gp_model(torch.cat([train_x, train_x]), torch.cat([y1_inds, y2_inds]))
+            output = gp_model(
+                torch.cat([train_x, train_x]), torch.cat([y1_inds, y2_inds])
+            )
             loss = -mll(output, torch.cat([train_y1, train_y2]))
             loss.backward()
             optimizer.step()
