+++ test/priors/test_multivariate_normal_prior.py
@@ -3,7 +3,7 @@ from __future__ import absolute_import, division, print_function, unicode_litera
 import unittest
 
 import torch
-from gpytorch.priors_torch import MultivariateNormalPrior
+from gpytorch.priors import MultivariateNormalPrior
 from torch.distributions import MultivariateNormal
