--- test/priors_torch/__init__.py
@@ -1,1 +1,0 @@
-from __future__ import absolute_import, division, print_function, unicode_literals
